requirements.txt
google-api-core
google-generativeai
ollama